+++ extensions/pins/CHANGELOG.md
@@ -1,10 +1,10 @@
 # Pins Changelog
 
-## [Minor Improvements] - 2023-05-20
+## [Minor Improvements] - 2023-05-22
 
 - Added "Preferred Browser" setting to allow you to choose which browser to open links in
 - Added support for tilde expansion in paths
-- Added menu bar menu items for "Pin This Tab" and "Pin This Directory"
+- Added menu items for "Pin This Tab", "Pin This Directory", and "Pin This App"
 - Added 'Show "Pin This" shortcut' setting to allow you to hide the "Pin This" items
 - Added "Create New Pin" and "Create New Group" actions
 - Added "Duplicate Pin" action